Refactor page components
