+++ packages/ui/src/App.tsx
@@ -7,7 +7,7 @@ import { useMockMembers } from './mocks/useMockMembers'
 import { useSudoBudget } from './mocks/useSudoBudget'
 import { MyAccounts } from './pages/Profile/MyAccounts'
 import { MyMemberships } from './pages/Profile/MyMemberships'
-import { GroupStorage } from './pages/WorkingGroups/GroupStorage'
+import { WorkingGroup } from './pages/WorkingGroups/WorkingGroup'
 import { WorkingGroups } from './pages/WorkingGroups/WorkingGroups'
 
 export function App() {
@@ -20,7 +20,7 @@ export function App() {
         <Route exact path="/profile" component={MyAccounts} />
         <Route exact path="/profile/memberships" component={MyMemberships} />
         <Route exact path="/groups" component={WorkingGroups} />
-        <Route exact path="/groups/grouppreview" component={GroupStorage} />
+        <Route exact path="/groups/grouppreview" component={WorkingGroup} />
         <Redirect exact from="/" to="/profile" />
       </Switch>
       <ConnectionStatus />
