+++ packages/ui/src/pages/Profile/MyAccounts/index.ts
@@ -1,0 +1,1 @@
+export * from './MyAccounts'
