--- packages/ui/src/pages/Profile/MyMemberships.tsx
@@ -1,34 +1,0 @@
-import React from 'react'
-
-import { Page } from '../../components/page/Page'
-import { PageContent } from '../../components/page/PageContent'
-import { PageHeader } from '../../components/page/PageHeader'
-import { PageTitle } from '../../components/page/PageTitle'
-import { SideBar } from '../../components/page/SideBar'
-import { Breadcrumbs } from '../../components/page/Sidebar/Breadcrumbs/Breadcrumbs'
-import { Memberships } from './MyMemberships/Memberships'
-import { MyProfileTabs } from './MyProfileTabs'
-import { MyProfile, MyMembershipsContent } from './Profile'
-
-export const MyMemberships = () => (
-  <Page>
-    <SideBar />
-    <PageContent>
-      <Breadcrumbs
-        crumbs={[
-          { href: '#', text: 'My Profile' },
-          { href: '#', text: 'My Memberships' },
-        ]}
-      />
-      <MyProfile>
-        <PageHeader>
-          <PageTitle>My profile</PageTitle>
-          <MyProfileTabs />
-        </PageHeader>
-        <MyMembershipsContent>
-          <Memberships />
-        </MyMembershipsContent>
-      </MyProfile>
-    </PageContent>
-  </Page>
-)
