+++ packages/ui/src/pages/Profile/MyMemberships/index.ts
@@ -1,0 +1,1 @@
+export * from './MyMemberships'
