+++ packages/ui/src/pages/WorkingGroups/WorkingGroup/WorkingGroup.tsx
@@ -1,18 +1,18 @@
 import React from 'react'
 
-import { Page } from '../../components/page/Page'
-import { PageContent, ContentWithSidepanel, SidePanel } from '../../components/page/PageContent'
-import { PageHeader } from '../../components/page/PageHeader'
-import { PageTitle } from '../../components/page/PageTitle'
-import { SideBar } from '../../components/page/SideBar'
-import { Breadcrumbs } from '../../components/page/Sidebar/Breadcrumbs/Breadcrumbs'
-import { Statistics } from '../../components/statistics/Stats'
-import { Label } from '../../components/typography'
-import { GroupContent } from './GroupsComponents'
-import { GroupStorageTabs } from './GroupStorageTabs'
-import { OpeningsCategories, OpeningsCategory, OpeningsList, OpeningItem } from './Openings/OpeningsComponents'
+import { Page } from '../../../components/page/Page'
+import { PageContent, ContentWithSidepanel, SidePanel } from '../../../components/page/PageContent'
+import { PageHeader } from '../../../components/page/PageHeader'
+import { PageTitle } from '../../../components/page/PageTitle'
+import { SideBar } from '../../../components/page/SideBar'
+import { Breadcrumbs } from '../../../components/page/Sidebar/Breadcrumbs/Breadcrumbs'
+import { Statistics } from '../../../components/statistics/Stats'
+import { Label } from '../../../components/typography'
+import { GroupContent } from '../Components'
+import { OpeningsCategories, OpeningsCategory, OpeningsList, OpeningItem } from '../Openings/OpeningsComponents'
+import { WorkingGroupTabs } from './WorkingGroupTabs'
 
-export function GroupStorage() {
+export function WorkingGroup() {
   return (
     <Page>
       <SideBar />
@@ -26,7 +26,7 @@ export function GroupStorage() {
         />
         <PageHeader>
           <PageTitle>Storage</PageTitle>
-          <GroupStorageTabs />
+          <WorkingGroupTabs />
           <ContentWithSidepanel>
             <GroupContent>
               <Statistics
