+++ packages/ui/src/pages/WorkingGroups/WorkingGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './WorkingGroup'
