+++ packages/ui/src/pages/WorkingGroups/WorkingGroups/index.ts
@@ -1,0 +1,1 @@
+export * from './WorkingGroups'
